Display movies response for specifig genre

+++ app/src/main/java/com/example/jvmori/discovermovies/data/network/response/MovieResult.kt
@@ -28,7 +28,7 @@ data class MovieResult(
     @SerializedName("video")
     val video: Boolean,
     @SerializedName("vote_average")
-    val voteAverage: Int,
+    val voteAverage: Double,
     @SerializedName("vote_count")
-    val voteCount: Int
+    val voteCount: Double
 )
